common code fixes to support ios; C#-like approach to data stream reading
