+++ hellomap-android/Extensions.cs
@@ -7,23 +7,6 @@ namespace CartoMobileSample
 {
 	static class Extensions
 	{
-		public static string[] ToStringArray(this List<Type> list)
-		{
-			string[] sampleNames = new string[list.Count];
-
-			for (int i = 0; i < list.Count; i++)
-			{
-				sampleNames[i] = list[i].Name;
-			}
-
-			return sampleNames;
-		}
-
-		public static ICharSequence ToCharSequence(this string text)
-		{
-			return new Java.Lang.String(text);
-		}
-
 		public static string GetTitle(this Type type)
 		{
 			return type.Name.Replace("Activity", "");
@@ -42,22 +25,9 @@ namespace CartoMobileSample
 			}
 		}
 
-		public static long ToLong(this ulong ulongValue)
+		public static ICharSequence ToCharSequence(this string text)
 		{
-			return (long)ulongValue;
-		}
-
-		public static string ConvertFromSecondsToHours(this double sec)
-		{
-			int hours = ((int)sec) / 3600,
-
-			remainder = ((int)sec) % 3600,
-			minutes = remainder / 60,
-			seconds = remainder % 60;
-
-			return ((hours < 10 ? "0" : "") + hours
-				+ "h" + (minutes < 10 ? "0" : "") + minutes
-				+ "m" + (seconds < 10 ? "0" : "") + seconds + "s");
+			return new Java.Lang.String(text);
 		}
 
 		public static void MakeToast(this Android.App.Activity activity, string message)
