+++ hellomap-ios/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="CartoMobileSDK" version="4.0.0-alphapreview1" targetFramework="xamarinios10" />
+  <package id="CartoMobileSDK" version="4.0.0-build27" targetFramework="xamarinios10" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="xamarinios10" />
 </packages>
